requirements.txt
Flask
gunicorn
openstacksdk==0.99.0
os_client_config
pyyaml